requirements.txt
flask
pyserial
numpy
scipy